requirements.txt
black
fsspec==2022.10.0
functions-framework==3.3.0
gcsfs==2022.10.0
google-cloud-storage==2.5.0
jupyter==1.0.0
matplotlib
mlflow==1.30.0
multiprocess==0.70.14
mypy==1.0.1
numpy==1.23.1
pandas==1.4.3
plotly==5.11.0
pylint==2.16.2
pytest==7.2.0
pytest-cov==4.0.0
pytest-mock==3.10.0
requests==2.25.0
scikit-learn==1.1.2
scipy==1.9.0
seaborn
